+++ R/relop_drop_columns.R
@@ -19,6 +19,7 @@ ex_data_table.relop_drop_columns <- function(optree,
                                              source_limit = NULL,
                                              source_usage = NULL,
                                              env = parent.frame()) {
+  force(env)
   wrapr::stop_if_dot_args(substitute(list(...)), "rqdatatable::ex_data_table.relop_drop_columns")
   cols <- optree$columns
   if(is.null(source_usage)) {