+++ R/relop_natural_join.R
@@ -38,6 +38,7 @@ ex_data_table.relop_natural_join <- function(optree,
                                              source_usage = NULL,
                                              source_limit = NULL,
                                              env = parent.frame()) {
+  force(env)
   wrapr::stop_if_dot_args(substitute(list(...)), "rqdatatable::ex_data_table.relop_natural_join")
   if(is.null(source_usage)) {
     source_usage <- columns_used(optree)