+++ R/relop_non_sql.R
@@ -254,6 +254,7 @@ ex_data_table.relop_non_sql <- function(optree,
                                         source_usage = NULL,
                                         source_limit = NULL,
                                         env = parent.frame()) {
+  force(env)
   wrapr::stop_if_dot_args(substitute(list(...)), "rqdatatable::ex_data_table.relop_non_sql")
   if(is.null(source_usage)) {
     source_usage <- columns_used(optree)