+++ R/relop_null_replace.R
@@ -24,6 +24,7 @@ ex_data_table.relop_null_replace <- function(optree,
                                              source_usage = NULL,
                                              source_limit = NULL,
                                              env = parent.frame()) {
+  force(env)
   wrapr::stop_if_dot_args(substitute(list(...)), "rqdatatable::ex_data_table.relop_null_replace")
   if(is.null(source_usage)) {
     source_usage <- columns_used(optree)