+++ R/relop_orderby.R
@@ -39,6 +39,7 @@ ex_data_table.relop_orderby <- function(optree,
                                         source_usage = NULL,
                                         source_limit = NULL,
                                         env = parent.frame()) {
+  force(env)
   wrapr::stop_if_dot_args(substitute(list(...)), "rqdatatable::ex_data_table.relop_orderby")
   if(is.null(source_usage)) {
     source_usage <- columns_used(optree)