+++ R/relop_select_rows.R
@@ -23,6 +23,7 @@ ex_data_table.relop_select_rows <- function(optree,
                                             source_usage = NULL,
                                             source_limit = NULL,
                                             env = parent.frame()) {
+  force(env)
   wrapr::stop_if_dot_args(substitute(list(...)), "rqdatatable::ex_data_table.relop_select_rows")
   if(is.null(source_usage)) {
     source_usage <- columns_used(optree)