+++ R/relop_set_indicator.R
@@ -26,6 +26,7 @@ ex_data_table.relop_set_indicator <- function(optree,
                                               source_usage = NULL,
                                               source_limit = NULL,
                                               env = parent.frame()) {
+  force(env)
   wrapr::stop_if_dot_args(substitute(list(...)), "rqdatatable::ex_data_table.relop_set_indicator")
   if(is.null(source_usage)) {
     source_usage <- columns_used(optree)