+++ NEWS.md
@@ -1,4 +1,8 @@
 
+# rqdatatable 1.0.1 2018/09/18
+
+  * force parent.frame
+
 # rqdatatable 1.0.0 2018/09/10
 
   * allow no group columns project.